Move IOFS stuff out into separate file, add basic tmpfs implementation for linux only
