+++ api/agent/agent.go
@@ -6,10 +6,8 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
-	"os"
 	"strings"
 	"sync"
 	"time"
@@ -1159,77 +1157,6 @@ type container struct {
 	stats  *drivers.Stats
 }
 
-// IOFS object encapsulates a temporary path and the method to clean it up
-type iofs interface {
-	io.Closer
-	AgentPath() string
-	DockerPath() string
-}
-
-type noopIOFS struct {
-}
-
-func (n *noopIOFS) AgentPath() string {
-	return ""
-}
-
-func (n *noopIOFS) DockerPath() string {
-	return ""
-}
-
-func (n *noopIOFS) Close() error {
-	return nil
-}
-
-type directoryIOFS struct {
-	agentPath  string
-	dockerPath string
-}
-
-func (d *directoryIOFS) AgentPath() string {
-	return d.agentPath
-}
-
-func (d *directoryIOFS) DockerPath() string {
-	return d.dockerPath
-}
-
-func (d *directoryIOFS) Close() error {
-	err := os.RemoveAll(d.agentPath)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func newDirectoryIOFS(cfg *Config) (*directoryIOFS, error) {
-	// XXX(reed): need to ensure these are cleaned up if any of these ops in here fail...
-
-	dir := cfg.IOFSAgentPath
-	if dir == "" {
-		// /tmp should be a memory backed filesystem, where we can get user perms
-		// on the socket file (fdks must give write permissions to users on sock).
-		// /var/run is root only, hence this...
-		dir = "/tmp"
-	}
-
-	// create a tmpdir
-	iofsAgentDir, err := ioutil.TempDir(dir, "iofs")
-	if err != nil {
-		return nil, fmt.Errorf("cannot create tmpdir for iofs: %v", err)
-	}
-
-	iofsRelPath, err := filepath.Rel(dir, iofsAgentDir)
-	if err != nil {
-		return nil, fmt.Errorf("cannot relativise iofs path: %v", err)
-	}
-	iofsDockerDir := filepath.Join(cfg.IOFSMountRoot, iofsRelPath)
-
-	return &directoryIOFS{iofsAgentDir, iofsDockerDir}, nil
-}
-
-var _ iofs = &directoryIOFS{}
-
 //newHotContainer creates a container that can be used for multiple sequential events
 func newHotContainer(ctx context.Context, call *call, cfg *Config) (*container, error) {
 	// if freezer is enabled, be consistent with freezer behavior and
@@ -1281,7 +1208,7 @@ func newHotContainer(ctx context.Context, call *call, cfg *Config) (*container, 
 	if call.Format == models.FormatHTTPStream {
 		// XXX(reed): we should also point stdout to stderr, and not have stdin
 
-		iofs, err = newDirectoryIOFS(cfg)
+		iofs, err = newTmpfsIOFS(cfg)
 		if err != nil {
 			return nil, err
 		}
